requirements.txt
torch>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
matplotlib>=3.7.0
pyyaml>=6.0
tiktoken>=0.5.0
wandb>=0.15.0
pytest>=7.4.0
flash-attn>=2.0.0;platform_system=="Linux"
